server/requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.0
pydantic>=2.8.0
openenv-core>=0.1.0

